transport: Use just a set instead of tuple as key in variation map
The size argument was there as a bad way of soriting map in descending order
(therefore the `size = ((size_t)-1) - filter_ids.size()`).

The only usage of variation map is the "distrubute" method.
And there we don't use that first key argument "size".

Therefore we can just use simple set-to-vector map with custom
comparator to mimic the tuple {size, set} behavior