Remove redundant segment size check
We don't need to perform this check. If a segment is too small,
no packets will fit, but that is a configuration issue, not a code issue.
It simply means the user has specified an insufficient segment size in
the configuration. There's no need to crash in this case.

Instead, after rotation, we get a fresh segment without any packets.
We should then check whether its size is sufficient to fit the current packet.