Revert changes that made tcpdump print to stdout instead of a file
This is a consolidation of two revert commits:

Revert "Optimize tcpdump pcap output for redirected stdout"
This reverts commit 2e076d26f7de8993e7d930d7447e383604f46c96.

Revert "Write PCAP data to a CLI-passed socket using SCM_RIGHTS, not disk files"
This reverts commit 277c405a541fdf3dc45efb506a888a77a94a4141.

We want to implement two modes for yanet-cli tcpdump command: read and
follow. The old behavior of storing packets in shm and an ability to create
pcap files should be the "read" mode.

This change is made to preserve history, instead of simply removing commits.
Them could be useful later